requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
numpy
httpx
python-dotenv
scipy==1.11.4
tf-keras
PyMuPDF
torch
torchvision
torchaudio
unstructured[pdf,local-inference]
rank_bm25